moved common helper funcs to range_op/utils and polished funcs
