+++ magi_attention/common/range_op/_range_fill.py
@@ -19,6 +19,8 @@ import triton.language as tl
 
 from magi_attention.utils import nvtx
 
+from .utils import _calc_cu_range_sizes, _calc_ranges_row_map
+
 __all__ = ["range_fill_"]
 
 
@@ -63,23 +65,23 @@ def range_fill_kernel(
 def range_fill_(
     input: torch.Tensor,
     ranges: torch.Tensor,
-    cu_range_sizes: torch.Tensor,
-    total_size: int,
     val: float,
     dim: int = 0,
-    row_map: torch.Tensor | None = None,
+    **kwargs,
 ) -> torch.Tensor:
     """
     Fill specified ranges in the input tensor with a given value.
 
     Args:
-        input: Tensor to be filled in-place
-        ranges: Tensor of [start, end] ranges to fill
-        cu_range_sizes: Cumulative sizes of ranges
-        total_size: Total number of rows to process
+        input (torch.Tensor): Tensor to be filled in-place
+        ranges (torch.Tensor): Tensor of [start, end] ranges to fill
         val: Value to fill the ranges with
         dim: Dimension along which to perform the fill operation
-        row_map: Optional mapping from row indices to range indices
+
+        kwargs:
+            - cu_range_sizes (torch.Tensor): Cumulative sizes of ranges
+            - total_size (int): Total number of rows to process
+            - row_map (torch.Tensor): mapping from row indices to range indices
 
     Returns:
         The in-place filled input tensor
@@ -88,12 +90,34 @@ def range_fill_(
     # Check that input has no gradient
     assert not input.requires_grad, "input must not require grad"
 
-    # Get the number of ranges
-    n_ranges = ranges.shape[0]
+    # ---   calculate meta   --- #
 
     # Return directly if empty tensor
-    if n_ranges == 0 or input.numel() == 0:
+    if ranges.shape[0] == 0 or input.numel() == 0:
         return input
+
+    # Make ranges contiguous
+    ranges = ranges.contiguous()
+
+    # Calculate cu_range_sizes and total_size if not provided
+    cu_range_sizes = kwargs.pop("cu_range_sizes", None)
+    total_size = kwargs.pop("total_size", None)
+    if cu_range_sizes is None or total_size is None:
+        cu_range_sizes, total_size = _calc_cu_range_sizes(
+            ranges,
+            device=input.device,
+        )
+    else:
+        cu_range_sizes = cu_range_sizes.contiguous()
+
+    # Calculate row_map if not provided
+    row_map = kwargs.pop("row_map", None)
+    if row_map is None:
+        row_map = _calc_ranges_row_map(ranges, total_size)
+    else:
+        row_map = row_map.contiguous()
+
+    # ---   pre-process input/output   --- #
 
     # Handle the case when dim is not 0
     if dim != 0:
@@ -101,18 +125,10 @@ def range_fill_(
     else:
         kernel_input = input.contiguous()
 
-    ranges = ranges.contiguous()
-    cu_range_sizes = cu_range_sizes.contiguous()
-
     # Calculate stride (considering memory step size of elements)
     input_stride = kernel_input.stride(0)
 
-    if row_map is None:
-        row_map = torch.arange(0, ranges.shape[0], device=ranges.device)
-        range_sizes = ranges[:, 1] - ranges[:, 0]
-        row_map = torch.repeat_interleave(
-            row_map, range_sizes, dim=0, output_size=total_size
-        )
+    # ---   calculate grid size   --- #
 
     M = total_size
     N = kernel_input.numel() // kernel_input.shape[0]
@@ -122,6 +138,8 @@ def range_fill_(
 
     # Calculate grid size
     grid = (M, N_BLOCK)
+
+    # ---   launch kernel   --- #
 
     # Launch kernel
     range_fill_kernel[grid](
@@ -135,6 +153,8 @@ def range_fill_(
         N_BLOCK,
         ELEM_PER_BLOCK,
     )
+
+    # ---   post-process output   --- #
 
     # If transposed earlier, transpose back
     if dim != 0:
